type: 타입분리
